fix: prefer background-color property (#206)
As described in the issue, using the `background` CSS property to set
only the `background-color` makes it more difficult to set other
individual `background-*` properties downstream, since they get
clobbered unless the selector chain is specific enough.

As a rule, it would be better to use the `background-color` property
specifically when that is all that is being set. Per @lex111's advice,
this commit only updates the example I am having problems with.